refactor navigate to logout hook
